documenting everything and clearing up unused stuff
