+++ app/controllers/metrics_controller.rb
@@ -1,10 +1,12 @@
+# Controller for Metric actions
 class MetricsController < ApplicationController
-  #layout "metric"
-  #TODO - why is this here?
+
+# The show action for a single Metric
   def show
     @metric = Metric.get(params[:id])
   end
 
+# The index action for showing all metrics
   def index
     @metrics = Metric.all.page(params[:page])
   end
