--- app/helpers/comparison_helper.rb
@@ -1,2 +1,0 @@
-module ComparisonHelper
-end
