--- app/helpers/comparisons_helper.rb
@@ -1,2 +1,0 @@
-module ComparisonsHelper
-end
