--- app/helpers/graph_helper.rb
@@ -1,2 +1,0 @@
-module GraphHelper
-end
