--- app/helpers/graphs_helper.rb
@@ -1,2 +1,0 @@
-module GraphsHelper
-end
