--- app/helpers/metric_helper.rb
@@ -1,2 +1,0 @@
-module MetricHelper
-end
