--- app/helpers/metrics_helper.rb
@@ -1,2 +1,0 @@
-module MetricsHelper
-end
