--- app/helpers/msrun_helper.rb
@@ -1,2 +1,0 @@
-module MsrunHelper
-end
