+++ app/views/layouts/_header.html.haml
@@ -1,4 +1,3 @@
-- categories = ["uplc", "chromatography", "ms1", "dynamic_sampling", "ion_source", "ion_treatment", "peptide_ids", "ms2", "run_comparison"]
 %header
   %nav.round
     %ul.header
@@ -10,7 +9,7 @@
     %ul.subheader
       - if ['comparisons'].include?(params[:controller]) and params[:id]
         %li= link_to "Comparison Home", comparison_path(@comparison)# + '/' + params[:id]
-        - categories.each do |cat|
+        - CATEGORIES.each do |cat|
           %li= link_to "#{cat.humanize}", "#{comparison_path(@comparison)}/#{cat.downcase}"
   = link_to logo, root_path
   - if flash[:error]
