+++ lib/config.rb
@@ -20,6 +20,7 @@ require 'yaml'
 File.open("archiver_defaults.yml", 'w') {|out| out.print App_defaults.to_yaml }
 # These are the config settings for the app, as generated by deeper_merging the defaults with the settings
 AppConfig = App_defaults.deeper_merge(YAML.load_file('archiver_config.yml'))
+# This constant contains an array of all the root storage locations
 Roots = AppConfig[:nodes].map {|h| h.last[:archive_root]}
 # Default settings for the QC automatic processing
 Qc_defaults = {
