+++ todo.markdown
@@ -7,6 +7,7 @@
 *	Message Passing # TEST IT and the accompanying metric starting...
 * Message clearing fxns so that I can clean out the lists... maybe monthly?
 * Implement database updating of locations and the file moving system
+* Convert the Knowntrends hash to one more like the QC defaults one... where each object can be set individually
 
 ###Website
   - Tracking of what metric comparisons are already produced and available for viewing (Ensure they can't be recreated)
@@ -15,7 +16,6 @@
   - Add breadcrumbs for graph heirarchy?
       -- might be had by the simple "up" link I've added.
   - figure out how to AJAX update the alerts page  
-  - Colored arrows defining the positive or negative trends when known for each metric  
   - Make a special viewing page for any graphs that are significantly different (T-test) __????  Would this really help???__
     * Or a page of links to the images?
   * http://www.datatables.net/ ????  Would this make that table easier?
